requirements.txt
decord==0.6.0
gpytorch==1.9.1
h5py==2.10.0
json5==0.9.6
matplotlib==3.5.1
numpy==1.22.4
opencv-python==4.2.0.34
Pillow==9.0.0
protobuf==3.20.3
PyYAML==6.0
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
tensorboard==2.8.0
tensorboardX==2.0
torch==1.13.1
torchvision==0.9.1
tqdm==4.62.3